requirements.txt
ArchivesSnake==0.8.1
attrs==19.1.0
boltons==19.1.0
certifi==2019.3.9
chardet==3.0.4
clinner==1.12.3
colorlog==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.13
django-js-asset==1.2.2
django-mptt==0.10.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
drf-yasg==1.15.0
gitdb2==2.0.5
GitPython==2.1.11
health-check==3.4.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
more-itertools==7.0.0
psycopg2-binary==2.8.2
PyJWT==1.7.1
pyrsistent==0.15.2
pytz==2019.1
PyYAML==5.1
requests==2.21.0
ruamel.yaml==0.15.94
six==1.12.0
smmap2==2.0.5
structlog==19.1.0
swagger-spec-validator==2.4.3
uritemplate==3.0.0
urllib3==1.24.3
